JsonType 수정
Double 타입이 아닌 Int 타입으로 변경

+++ JSONParser/JSONParser/JsonType.swift
@@ -19,7 +19,7 @@ protocol JsonType {
     var typeDescription: String { get }
 }
 
-extension Double: JsonType {
+extension Int: JsonType {
     var typeDescription: String {
         return TypeName.number.rawValue
     }
